requirements.txt
asgiref==3.8.1
Django==5.0
django-cors-headers==4.7.0
django-jsonfield-backport==1.0.5
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
graphql-core==3.2.6
PyJWT==2.9.0
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
strawberry-graphql==0.214.0
strawberry-graphql-django==0.29.0
typing_extensions==4.14.0
tzdata==2025.2
